+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - `CompletionCandidates.Fixed` now has a secondary convenience constructor that take a `vararg` of `String`s
+- `CompletionCadidates.Custom`, which allows you to call other binaries or write a script to generate completions.
 
 ### Changed
 - `option()` and `argument()` now take optional `completionCandidates` parameters to override how completion is generated. The constructor and `copy` functions of `OptionsWithValues` and `ProcessedArgument` have changed to support default values.